Add the very first IaC Scanner docs
In this commit we are paving the way for establishing the initial
documentation for IaC Scanner. The documentation is currently in
the form of .rst files and is prepared to be build with Read the
Docs and Sphinx documentation tool generator. With these changes
we set the structure for hosting our docs on GitHub Pages.